break Map into three files
